Add TERMINAL macro defining terminal emulator app, default to foot
Previously, the default terminal program was footie, a foot wrapper script that ensured there is a foot server socket where terminal emulator instances (clients) would connect to. This has the advantage of lower resource usage and lower startup overhead but unfortunately spawning a new terminal within the same directory using MOD+space does not reread the config and breaks the seamlessness of dynamic theming. Perhaps in the future, foot server would implement readreading the config file forcibly, by say sending a signal in which case it would fix all issues there are with this approach as we could call this forceful rereading inside themeset and that would be an overall win. For now though, users with lower end specs can change this value to footie to reduce the overhead again. Remember also of setting the terminal emulator variable in .zprofile as well.